write scanner and add etl

+++ main/include/drivers/distanceSensor/HC_SR04_DistanceSensor.h
@@ -1,14 +1,14 @@
 #pragma once
 
 #include "IDistanceSensor.h"
-#include "externals/gpio/IGPIO.h"
+#include "drivers/gpio/IGPIO.h"
 
 class HC_SR04_DistanceSensor : public IDistanceSensor {
    public:
     HC_SR04_DistanceSensor(IGPIO &echo, IGPIO &trigger);
     ~HC_SR04_DistanceSensor();
 
-    uint64_t calculateDistanceCm() override;
+    uint16_t calculateDistanceCm() override;
 
    private:
     IGPIO &_echo;
